+++ src/app/about/page.tsx
@@ -4,7 +4,7 @@ import Mission from './components/Mission'
 import Vission from './components/Vission'
 import Stats from '@/components/Stats'
 import WhyUs from './components/WhyUs'
-import HowItWorks from './components/HowItWorks'
+import HowItWorks from '../../components/HowItWorks'
 import Team from './components/Team'
 import PartnersCard from './components/PartnersCard'
 import Newletter from './components/Newletter'
